Move the function out of rwlock
